+++ src/components/index.ts
@@ -6,10 +6,12 @@ export { Input } from './Input';
 export { LoginForm } from './LoginForm/LoginForm';
 export { LogoutButton } from './LogoutButton/LogoutButton';
 export { MainLayout } from './MainLayout';
+export { MobileSidebar } from './MobileSidebar/MobileSidebar';
 export { Pagination } from './Pagination';
 export { ProductCard } from './ProductCard';
 export { ProductFiltersBar } from './ProductsFilters/ProductsFiltersBar';
 export { ProtectedRoute } from './ProtectedRoute/ProtectedRoute';
 export { SearchInput } from './SearchInput/SearchInput';
+export { Sidebar } from './Sidebar/Sidebar';
 export { TableProducts } from './TableProducts/TableProducts';
 export { TextArea } from './TextArea';
